working on opening the expanded chat

+++ app/src/main/java/com/andrei/engine/repositoryInterfaces/BookingRepoInterface.kt
@@ -8,5 +8,5 @@ import retrofit2.http.Path
 
 interface BookingRepoInterface {
 @GET("/users/{userID}/bookings")
-fun getBookings(@Path("userID") userID:Int): CallWrapper<List<BookingDTO>>
+fun getBookings(@Path("userID") userID:Long): CallWrapper<List<BookingDTO>>
 }
