+++ app/src/main/java/com/andrei/engine/requestModels/NewBookingRequestModel.kt
@@ -12,5 +12,5 @@ data class NewBookingRequestModel(
         @SerializedName("carID")
         val carID:Long,
         @SerializedName("userID")
-        val userID:Int
+        val userID:Long
 )
